requirements.txt
asgiref==3.3.1
autopep8==1.5.5
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
firebase-admin==4.5.2
fuzzywuzzy==0.18.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
msgpack==1.0.2
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
xmltodict==0.12.0
